requirements.txt
openai
beautifulsoup4
discord
tqdm
boto3
python-dotenv